Add a namespace for AYTQ
The app is responsible for serving both AYTQ and Check Records.

A namespace for Check Records already exists but we don't have a
corresponding one for AYTQ.

I moved the sign in routes as well, which meant there needed to be
modifications to the tests to support multiple domains.
