+++ .env.test
@@ -1,6 +1,6 @@
 GOVUK_NOTIFY_API_KEY=override-locally
-CHECK_RECORDS_DOMAIN=http://127.0.0.1
-HOSTING_DOMAIN=http://127.0.0.1
+CHECK_RECORDS_DOMAIN=http://check_records.localhost
+HOSTING_DOMAIN=http://qualifications.localhost
 HOSTING_ENVIRONMENT=local
 IDENTITY_API_DOMAIN=override-locally
 IDENTITY_CLIENT_ID=override-locally
