--- app/controllers/identity_users_controller.rb
@@ -1,4 +1,0 @@
-class IdentityUsersController < QualificationsInterfaceController
-  def show
-  end
-end
