+++ app/controllers/qualifications/identity_users_controller.rb
@@ -1,0 +1,6 @@
+module Qualifications
+  class IdentityUsersController < QualificationsInterfaceController
+    def show
+    end
+  end
+end
