+++ app/controllers/qualifications/static_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Qualifications
+  class StaticController < ApplicationController
+    layout "two_thirds"
+  end
+end
