--- app/controllers/static_controller.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class StaticController < ApplicationController
-  layout "two_thirds"
-end
