--- app/controllers/users/omniauth_callbacks_controller.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-class Users::OmniauthCallbacksController < ApplicationController
-  def identity
-    auth = request.env["omniauth.auth"]
-    @user = User.from_identity(auth)
-    session[:identity_user_id] = @user.id
-    session[:identity_user_token] = auth.credentials.token
-    session[:identity_user_token_expiry] = auth.credentials.expires_at
-
-    log_auth_credentials_in_development(auth)
-    flash[:notice] = "Signed in successfully."
-    redirect_to qualifications_path
-  end
-
-  private
-
-  def log_auth_credentials_in_development(auth)
-    if Rails.env.development?
-      Rails.logger.debug auth.credentials.token
-      Rails.logger.debug Time.zone.at auth.credentials.expires_at
-    end
-  end
-end
