--- app/controllers/users/sign_out_controller.rb
@@ -1,7 +1,0 @@
-class Users::SignOutController < QualificationsInterfaceController
-  skip_before_action :handle_expired_token!
-
-  def new
-    session[:identity_user_id] = nil if user_signed_in?
-  end
-end
