+++ app/views/layouts/base.html.erb
@@ -58,13 +58,13 @@
 
           <ul class="govuk-footer__inline-list">
             <li class="govuk-footer__inline-list-item">
-              <%= govuk_link_to("Accessibility", "/accessibility", class: "govuk-footer__link") %>
+              <%= govuk_link_to("Accessibility", "/qualifications/accessibility", class: "govuk-footer__link") %>
             </li>
             <li class="govuk-footer__inline-list-item">
-              <%= govuk_link_to("Cookies", "/cookies", class: "govuk-footer__link") %>
+              <%= govuk_link_to("Cookies", "/qualifications/cookies", class: "govuk-footer__link") %>
             </li>
             <li class="govuk-footer__inline-list-item">
-              <%= govuk_link_to("Privacy notice", "/privacy", class: "govuk-footer__link") %>
+              <%= govuk_link_to("Privacy notice", "/qualifications/privacy", class: "govuk-footer__link") %>
             </li>
           </ul>
