+++ app/views/qualifications/qualifications/show.html.erb
@@ -31,7 +31,7 @@
           <strong><%= @teacher.trn %></strong>
         </p>
 
-        <%= govuk_link_to "View and update details", identity_user_path %>
+        <%= govuk_link_to "View and update details", qualifications_identity_user_path %>
       </div>
     </div>
   </div>
