+++ app/views/qualifications/starts/show.html.erb
@@ -17,7 +17,7 @@
       You’ll need to create a <strong>DfE Identity account</strong> to access your teaching qualifications. You'll use the account to sign-in to this service and other DfE teaching services.
     </p>
 
-    <%= button_to ["/users/auth/identity", "trn_token=#{params[:trn_token]}"].join('?'), class: "govuk-button govuk-button--start" do %>
+    <%= button_to ["/qualifications/users/auth/identity", "trn_token=#{params[:trn_token]}"].join('?'), class: "govuk-button govuk-button--start" do %>
       Start now
       <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
         <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z"></path>
