+++ app/views/qualifications/users/sign_in/new.html.erb
@@ -1,7 +1,7 @@
 <%=
   govuk_button_to(
     "Sign in with Identity",
-    "/users/auth/identity",
+    "/qualifications/users/auth/identity",
     method: :post
   )
 %>
