+++ app/views/qualifications/users/sign_out/new.html.erb
@@ -2,6 +2,6 @@
   <div class="govuk-grid-column-two-thirds">
     <h1 class="govuk-heading-xl">You are now signed out</h1>
 
-    <%= govuk_button_link_to "Back to GOV.UK", root_path  %>
+    <%= govuk_button_link_to "Back to GOV.UK", qualifications_root_path  %>
   </div>
 </div>
