+++ app/views/staff/mailer/invitation_instructions.text.erb
@@ -1,6 +1,6 @@
 <%= t("devise.mailer.invitation_instructions.hello", email: @resource.email) %>
 
-<%= t("devise.mailer.invitation_instructions.someone_invited_you", url: root_url) %>
+<%= t("devise.mailer.invitation_instructions.someone_invited_you", url: support_interface_staff_root_url) %>
 
 <%= accept_invitation_url(@resource, invitation_token: @token) %>
