+++ config/environments/development.rb
@@ -59,4 +59,5 @@ Rails.application.configure do
   # config.action_cable.disable_request_forgery_protection = true
   config.action_mailer.delivery_method = :notify
   config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
+  config.hosts = nil
 end
