+++ terraform/workspace_variables/preprod.tfvars.json
@@ -1,11 +1,11 @@
 {
-    "environment_name": "preprod",
-    "key_vault_name": "s165t01-aytq-pp-kv",
-    "resource_group_name": "s165t01-aytq-pp-rg",
-    "storage_account_name": "s165t01aytqtfstatepp",
-    "enable_blue_green": true,
-    "app_service_plan_sku": "S1",
-    "keyvault_logging_enabled": true,
-    "storage_log_categories": [],
-    "resource_prefix": "s165t01-aytq"
+  "environment_name": "preprod",
+  "key_vault_name": "s165t01-aytq-pp-kv",
+  "resource_group_name": "s165t01-aytq-pp-rg",
+  "storage_account_name": "s165t01aytqtfstatepp",
+  "enable_blue_green": true,
+  "app_service_plan_sku": "S1",
+  "keyvault_logging_enabled": true,
+  "storage_log_categories": [],
+  "resource_prefix": "s165t01-aytq"
 }
