+++ terraform/workspace_variables/production.tfvars.json
@@ -1,14 +1,14 @@
 {
-    "environment_name": "production",
-    "key_vault_name": "s165p01-aytq-pd-kv",
-    "resource_group_name": "s165p01-aytq-pd-rg",
-    "storage_account_name": "s165p01aytqtfstatepd",
-    "enable_blue_green": true,
-    "app_service_plan_sku": "P1v2",
-    "keyvault_logging_enabled": true,
-    "storage_log_categories": [],
-    "resource_prefix": "s165p01-aytq",
-    "worker_count": 3,
-    "postgres_flexible_server_sku": "GP_Standard_D2s_v3",
-    "enable_postgres_high_availability": true
+  "environment_name": "production",
+  "key_vault_name": "s165p01-aytq-pd-kv",
+  "resource_group_name": "s165p01-aytq-pd-rg",
+  "storage_account_name": "s165p01aytqtfstatepd",
+  "enable_blue_green": true,
+  "app_service_plan_sku": "P1v2",
+  "keyvault_logging_enabled": true,
+  "storage_log_categories": [],
+  "resource_prefix": "s165p01-aytq",
+  "worker_count": 3,
+  "postgres_flexible_server_sku": "GP_Standard_D2s_v3",
+  "enable_postgres_high_availability": true
 }
